Move history code into its own file
